pyproject.toml
[project]
name = "skn21-final-5team"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.0",
    "bcrypt==4.1.2",
    "bitsandbytes>=0.47.0; platform_system != 'Darwin'",
    "cryptography>=46.0.4",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "fastembed>=0.7.4",
    "flashrank>=0.2.10",
    "itsdangerous>=2.2.0",
    "langchain-core>=1.2.8",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "openai>=2.16.0",
    "orjson>=3.10.18",
    "pandas>=3.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "redis>=5.2.1",
    "qdrant-client>=1.16.2",
    "sqlalchemy>=2.0.46",
    "torch>=2.7.0",
    "tqdm>=4.67.3",
    "transformers>=4.57.0",
    "uvicorn>=0.40.0",
    "python-jose[cryptography]>=3.3.0",
    "mcp>=1.26.0",
    "authlib>=1.6.9",
    "faiss-cpu>=1.13.2",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "oracledb>=3.4.2",
    "werkzeug>=3.1.6",
    "selenium>=4.41.0",
    "slack-sdk>=3.36.0",
    "curl-cffi>=0.14.0",
    "vite>=1.5.2",
    "django>=6.0.3",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
